+++ spec/pubmedly/pubmedly_spec.rb
@@ -5,5 +5,10 @@ module Pubmedly
     it "has a version number" do
       expect(VERSION).not_to be_nil
     end
+
+    it "defines field tags" do
+      expect(FIELD_TAGS).to be_a(Hash)
+      expect(FIELD_TAGS["Author"]).to eq ["au"]
+    end
   end
 end